Separate `Linkage` from `Function` module
