+++ crates/ir/src/function.rs
@@ -1,9 +1,7 @@
-use std::fmt;
-
 use fxhash::FxHashMap;
 use smallvec::SmallVec;
 
-use crate::module::ModuleCtx;
+use crate::{module::ModuleCtx, Linkage};
 
 use super::{module::FuncRef, DataFlowGraph, Layout, Type, Value};
 
@@ -87,28 +85,5 @@ impl Signature {
     #[doc(hidden)]
     pub fn set_ret_ty(&mut self, ty: Type) {
         self.ret_ty = ty;
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-/// Linkage of symbols.
-pub enum Linkage {
-    /// The symbol is defined in the module, and can be used from the outside of the module.
-    Public,
-
-    /// The symbol is defined in the module, and can NOT be called from another module.
-    Private,
-
-    /// The symbol is defined outside of the module.
-    External,
-}
-
-impl fmt::Display for Linkage {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> fmt::Result {
-        match self {
-            Self::Public => write!(f, "public"),
-            Self::Private => write!(f, "private"),
-            Self::External => write!(f, "external"),
-        }
     }
 }
