+++ crates/ir/src/lib.rs
@@ -7,6 +7,7 @@ pub mod insn;
 pub mod ir_writer;
 pub mod isa;
 pub mod layout;
+pub mod linkage;
 pub mod module;
 pub mod types;
 pub mod value;
@@ -16,10 +17,11 @@ mod bigint;
 pub use bigint::{I256, U256};
 pub use builder::Variable;
 pub use dfg::{Block, BlockData, DataFlowGraph};
-pub use function::{Function, Linkage, Signature};
+pub use function::{Function, Signature};
 pub use global_variable::{GlobalVariable, GlobalVariableData};
 pub use insn::{BranchInfo, DataLocationKind, Insn, InsnData};
 pub use layout::Layout;
+pub use linkage::Linkage;
 pub use module::Module;
 pub use types::Type;
 pub use value::{Immediate, Value, ValueData};
